+++ rust/lib/src/lib.rs
@@ -12,7 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! Determines the type of a file with deep-learning.
+//! Determines file content types using AI.
+//!
+//! This library indirectly depends on the ONNX Runtime through the `ort` [crate][ort-crate]. The
+//! final user is responsible to make sure the ONNX Runtime is available. The `ort` crate provides
+//! [many options][ort-linking] in this regard. The simplest option is to enable the default [cargo
+//! features][ort-features] of the `ort` crate by adding the following dependency to the
+//! `Cargo.toml` file of the final binary:
+//!
+//! ```toml
+//! [dependencies]
+//! ort = "=2.0.0-rc.12"
+//! ```
 //!
 //! # Examples
 //!
@@ -30,8 +41,12 @@
 //! # Ok(())
 //! # }
 //! ```
+//!
+//! [ort-crate]: https://crates.io/crates/ort
+//! [ort-linking]: https://ort.pyke.io/setup/linking
+//! [ort-features]: https://ort.pyke.io/setup/cargo-features
 
-#![cfg_attr(feature = "_doc", feature(doc_auto_cfg))]
+#![cfg_attr(feature = "_doc", feature(doc_cfg))]
 
 pub use crate::builder::Builder;
 pub use crate::content::{ContentType, MODEL_MAJOR_VERSION, MODEL_NAME};